streamlit/requirements.txt
streamlit>=1.35.0
psycopg2-binary>=2.9.0
